crontabs/pyproject.toml

[tool.isort]
known_first_party = ["dbm", "langs", "libs"]


[tool.poetry]
name = "crontabs"
version = "0.1.0"
description = "Cron implementation in docker"
authors = ["sam <sam@home.com>"]


[tool.poetry.dependencies]
python = ">=3.9,<4.0"
aiomysql = "0.2.0"
redis = "4.2"
argparse = "1.4.0"
betterproto = "2.0.0b4"
cryptography = "41.0.7"
dynamic-yaml = "1.3.3"
fastapi = "0.100.0"
fastapi-mail = "1.4.1"
geoip2 = "*"
grpclib = "0.4.5"
httpx = "0.24.1"
ipwhois = "1.2.0"
mysql-connector-python = "8.0.32"
py-redis = "1.1.1"
pydantic-settings = "2.1.0"
pydantic = { version = "2.5.2", extras = ["email"] }
pytest = "7.4.0"
pytest-xdist = "2.5.0"
pytest-asyncio = "0.23.2"
python-dateutil = "2.8.2"
python-pidfile = "3.1.1"
pytz = "2023.3.post1"
requests = "2.25.1"
sqlalchemy = "2.0.18"
urllib3 = "1.26.5"
isort = "6.0.1"
flake8 = "5.0.4"
black = "23.7.0"
